+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] 
 ### Added
+### Changed
+### Fixed
+- [upstream] QR Codes are now responsive
+- The wheel display is mobile friendly now
+
+
+## [v1.0.0] 14 Feb 2025
+### Added
 - Made the wheel
 - Added the campaign setup page
 - Made the customer view page visuals